requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
asyncio
aiohttp
datadog
python-dateutil
youtube_dl
requests
psutil
rethinkdb
jishaku
python-socketio[asyncio_client]
colorama
